Split data tests into reader and writer
